backend/requirements.txt
Django==5.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.10
Pillow==11.1.0
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
django-ckeditor==6.7.1